+++ docs/releases/v2.x/v2.0.x/v2.0.0/release-notes.md
@@ -33,11 +33,23 @@ Released on TBD.
 [Updated Classes](#updated-classes)
 - [`WeightedElement`](#weightedelement)
 
+[Updated Interfaces](#updated-interfaces)
+- [`PaletteColor`](#palettecolor)
+
 ----
 
 # Updated Classes
 
 ## `WeightedElement`
+
+### Deprecated Properties
+- `value: Type`
+- `weight: number`
+
+### New Properties
+- `VALUE: Type`
+- `WEIGHT: number`
+- `DISCRIMINATOR: Discriminators.WEIGHTED_ELEMENT`
 
 ```typescript
 declare interface WeightedElement<Type> extends Discriminable {
@@ -66,6 +78,80 @@ declare interface WeightedElement<Type> extends Discriminable {
      * @inheritDoc
      */
     readonly DISCRIMINATOR: Discriminators.WEIGHTED_ELEMENT;
+}
+```
+
+[Table of Contents](#table-of-contents)
+
+----
+
+# Updated Interfaces
+
+## PaletteColor
+
+`RGB` and `HSL` components are now optional.
+
+```typescript
+/**
+ * A color to be used in a {@link Palette}.
+ *
+ * @category Palette
+ */
+declare interface PaletteColor extends Discriminable {
+    /**
+     * The hex string representation of the color (format: `#RRGGBB`).
+     */
+    readonly HEX: string;
+
+    /**
+     * The name of the color.
+     */
+    readonly NAME: string;
+
+    /**
+     * The RGB (red, green, blue) components of the color.
+     */
+    readonly RGB?: {
+        /**
+         * The red component (0-255).
+         */
+        readonly R: number;
+
+        /**
+         * The green component (0-255).
+         */
+        readonly G: number;
+
+        /**
+         * The blue component (0-255).
+         */
+        readonly B: number;
+    };
+
+    /**
+     * The HSL (hue, saturation, lightness) components of the color.
+     */
+    readonly HSL?: {
+        /**
+         * The hue component (0-360).
+         */
+        readonly H: number;
+
+        /**
+         * The saturation component (0-100).
+         */
+        readonly S: number;
+
+        /**
+         * The lightness component (0-100).
+         */
+        readonly L: number;
+    };
+
+    /**
+     * @inheritDoc
+     */
+    readonly DISCRIMINATOR: Discriminators.PALETTE_COLOR;
 }
 ```